+++ cypress/utils/table.js
@@ -89,4 +89,29 @@ function* cumulativeCombinations(arr, current = []) {
   }
 }
 
-export { checkSorting, cypressApplyFilters, cumulativeCombinations };
+function selectRandomEnabledRows({
+  rows,
+  numberOfRowsToSelect,
+}) {
+  const enabledRows = Array.from(rows).filter(row => {
+    const checkbox = row.querySelector('input[type="checkbox"]');
+    if(!checkbox.hasAttribute('disabled')){
+      return true;
+    }
+  })
+  const rowCount = enabledRows.length
+
+  const randomIndices = [];
+  while (randomIndices.length < numberOfRowsToSelect) {
+    const randomIndex = Math.floor(Math.random() * rowCount);
+    if (!randomIndices.includes(randomIndex)) {
+      randomIndices.push(randomIndex);
+    }
+  }
+
+  randomIndices.forEach(index => {
+    enabledRows[index].querySelector('input[type="checkbox"]').click();
+  });
+}
+
+export { checkSorting, cypressApplyFilters, cumulativeCombinations, selectRandomEnabledRows };
